+++ include/PropertyAccessor/Generation/vpg_generation_option_property_accessor.hpp
@@ -20,7 +20,7 @@ class VPGGenerationOptionPropertyAccessor : public BasePropertyAccessor
 {
     PROPERTY_ACCESSOR_HEADER(bool, Bool)
     PROPERTY_ACCESSOR_HEADER(long, Long)
-    PROPERTY_ACCESSOR_HEADER(std::wstring, Wstring)
+    PROPERTY_ACCESSOR_HEADER(std::wstring, String)
     PROPERTY_ACCESSOR_OBJECT_HEADER(std::shared_ptr<IObject>, Object)
 
     PROPERTY_ACCESSOR_CONTAINER_HEADER